Move Language related functionality into config module
